Relocate metrics tracker to reporting domain
Move createMetricsTracker from src/shared/utils/metrics-tracker.js to src/shared/reporting/metrics-tracker.js so instrumentation helpers live alongside reporting barrels instead of the general utils directory.
