+++ src/shared/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "../utils/metrics-tracker.js";
+export * from "./metrics-tracker.js";
